ui: tag: make default tag label size more fitting of the font face
