+++ tag/tag.go
@@ -1,14 +1,9 @@
 package tag
 
 import (
-	"bufio"
-	"bytes"
 	"image"
-	"image/color"
 	"log"
 	"os"
-	"path/filepath"
-	"strings"
 	"sync"
 
 	"github.com/as/srv/fs"
@@ -17,8 +12,6 @@ import (
 	"github.com/as/edit"
 	"github.com/as/font"
 	"github.com/as/frame"
-	"github.com/as/path"
-	"github.com/as/text/find"
 	//	"github.com/as/ui/img"
 	"github.com/as/ui/win"
 	//"github.com/as/worm"
@@ -39,10 +32,6 @@ var (
 		},
 	}
 )
-
-func p(e mouse.Event) image.Point {
-	return image.Pt(int(e.X), int(e.Y))
-}
 
 type Tag struct {
 	Vis
@@ -77,18 +66,14 @@ func New(dev ui.Dev, conf *Config) *Tag {
 	return t
 }
 
-func (w *Tag) SetFont(ft font.Face) {
-	body := w.Body.(*win.Win)
-	if body == nil {
-		return
+func (t *Tag) Close() (err error) {
+	if t.Body != nil {
+		err = t.Body.Close()
 	}
-	if ft.Height() < 3 || w.Body == nil {
-		return
+	if t.Win != nil {
+		err = t.Win.Close()
 	}
-	body.SetFont(ft)
-	w.dirty = true
-	w.Mark()
-	body.Refresh()
+	return err
 }
 
 func (t *Tag) Dirty() bool {
@@ -100,84 +85,7 @@ func (t *Tag) Mark() {
 	t.Win.Mark()
 }
 
-func mustCompile(prog string) *edit.Command {
-	p, err := edit.Compile(prog)
-	if err != nil {
-		log.Printf("tag.go:/mustCompile/: failed to compile %q\n", prog)
-		return nil
-	}
-	return p
-}
-
-func (t *Tag) Open(basepath, title string) {
-	t.basedir = path.DirOf(basepath)
-	println(title)
-	t.Get(title)
-}
-
-func (t *Tag) Close() (err error) {
-	if t.Body != nil {
-		err = t.Body.Close()
-	}
-	if t.Win != nil {
-		err = t.Win.Close()
-	}
-	return err
-}
-
-func (t *Tag) Dir() string {
-	x := path.DirOf(t.FileName())
-	if IsAbs(x) {
-		return x
-	}
-	return filepath.Join(t.basedir, x)
-}
-
-func (t *Tag) fixtag(abs string) {
-	wtag := t.Win
-	p := wtag.Bytes()
-	maint := find.Find(p, 0, []byte{'|'})
-	if maint == -1 {
-		maint = int64(len(p))
-	}
-	wtag.Delete(0, maint+1)
-	wtag.InsertString(abs+"\tPut Del |", 0)
-	wtag.Refresh()
-}
-
-type GetEvent struct {
-	Basedir string
-	Name    string
-	Addr    string
-	IsDir   bool
-}
-
-func (t *Tag) abs() string {
-	name := t.FileName()
-	if !IsAbs(name) {
-		name = filepath.Join(t.basedir, name)
-	}
-	return name
-}
-
-func (t *Tag) FileName() string {
-	if t == nil || t.Win == nil {
-		return ""
-	}
-	name, err := bufio.NewReader(bytes.NewReader(t.Win.Bytes())).ReadString('\t')
-	if err != nil {
-		return ""
-	}
-	return strings.TrimSpace(name)
-}
-
-func pt(e mouse.Event) image.Point {
-	return image.Pt(int(e.X), int(e.Y))
-}
-
-var (
-	crimson = image.NewUniform(color.RGBA{70, 40, 56, 255})
-)
+//var crimson = image.NewUniform(color.RGBA{70, 40, 56, 255})
 
 func (t *Tag) Upload(wind screen.Window) {
 	var wg sync.WaitGroup
@@ -220,4 +128,19 @@ func (t *Tag) Refresh() {
 			wg.Done()
 		}()
 	}
+}
+
+func mustCompile(prog string) *edit.Command {
+	p, err := edit.Compile(prog)
+	if err != nil {
+		log.Printf("tag.go:/mustCompile/: failed to compile %q\n", prog)
+		return nil
+	}
+	return p
+}
+func p(e mouse.Event) image.Point {
+	return image.Pt(int(e.X), int(e.Y))
+}
+func pt(e mouse.Event) image.Point {
+	return image.Pt(int(e.X), int(e.Y))
 }
