+++ tag/tag_test.go
@@ -150,7 +150,7 @@ func testResize(t *testing.T, name string, tt *Tag) {
 		{0, 0, 10, 10, 0, 0, 10, 0, VisNone},
 		{0, 0, 11, 11, 0, 0, 11, 0, VisNone},
 		{0, 0, 12, 12, 0, 0, 12, 0, VisNone},
-		{0, 0, 22, 22, 0, 0, 22, 22, VisTag},
+		{0, 0, 22, 22, 0, 0, 22, 19, VisTag},
 		//		{0, 0, 23, 23, 0, 0, 23, 23, VisTag},
 		{0, 0, 1000, 1000, 0, 0, 1000, 1000, VisFull},
 		{0, 0, 500, 500, 0, 0, 500, 500, VisFull},
